Watches are scheduled on load. Unit tests needed.
The watches loaded from the watch store on load are now scheduled to run
and do send a message to run to the manager.

Tests are not yet done becasue I'm still having issues with the store
not properly starting because of a reliance on DETS. I need to figure
this out but don't have internet right now so I will once I land or I'll
come up with an intermediary solution.