Fix issue where adding watches wasn't working from UI
The UI expected a "timer_at" to always be available, so that has to be
defaulted to null.

Additionally, I changed the semantics of set_watch a bit so that it
returns a watch with more metadata, and returns a watch in general,
since before it just returned 'ok' which isn't very useful.