Don't overlap watch runs.
If you have a watch scheduled to run every 5000ms, but its taking
10000ms to run the watch, it doesn't make sense to continue backing
those ups in a queue. This ensures that only one job for any given watch
is running at any given time.

In the future, it'd be nice to change the state of the watch in some way
so that the user knows that this is happening.