GC/VM recycle for V8 VMs.
This is based on a "tick" system which is rather crude at the moment but
should work fine as we've seen this work in other production systems at
Kiip. The idea is that each watch is a tick in the VM and after a
certain number of ticks we trigger GC. Additionally, after another
number of ticks we recycle the entire VM justin case there are memory
leaks, since there historically have been in erlv8.