Lifeguard.get implementation is beginning. JS can turn into Erlang vals
Work on the Lifeguard.get method, which will retrieve values from a data
source written in Erlang. The implementation is beginning, and in its
current form it can turn the arguments into Erlang values. I'm currently
working on unit tests for this referentially transparent operation, as
well as improving it to handle recursive objects as well as circular
references cleanly.