Lifeguard.get returns data.
This involved creating some new methods to convert to and from Erlang/JS
values. These are all heavily unit tested.

Additionally, the interface for data sources changed slightly so that
they now take their ServerID in the form a first argument to start_link,
so they know what to register themselves as.