plugin manager can load plugins with enable_plugin
This currently will find the plugins in the configured plugin directory
and then add all the ebins to the code path, finally starting the
application.

In the future, there needs to be some more error checking here because
application:start doesn't start dependent applications and there will
undoubtedly be some.